ISC-2024-Skin-Cancer-Detection-with-3D-TBP/requirements.txt
#requirement.txt

#Install PyTorch Manually
#pip3 install torch==2.0.1+cu117 torchvision==0.15.2+cu117 -f https://download.pytorch.org/whl/torch_stable.html

#pip3 install -r requirements.txt

# torch and torchvision will be installed separately
pandas==2.0.3
h5py==3.8.0
Pillow==10.0.0
numpy==1.24.4
scikit-learn==1.3.0
tqdm==4.65.0
matplotlib==3.7.2
seaborn==0.12.2
albumentations==1.3.1


